django react project finished with tailwind

+++ client/src/pages/TaskFormPage.jsx
@@ -2,6 +2,7 @@ import { useEffect } from 'react'
 import {useForm} from 'react-hook-form'
 import { createTask, deleteTask, updateTask, getTask } from '../api/tasks.api'
 import {useNavigate, useParams} from 'react-router-dom'
+import { toast } from 'react-hot-toast'
 
 function TaskFormPage() {
 
@@ -14,8 +15,20 @@ function TaskFormPage() {
   const submitData = handleSubmit(async data => {
     if (params.id) {
       await updateTask(params.id, data)
+      toast.success('Task succesfully updated', {
+        style: {
+          background: "#101010",
+          color: "#fff"
+        }
+      })
     } else {
       await createTask(data)
+      toast.success('Task succesfully created', {
+        style: {
+          background: "#101010",
+          color: "#fff"
+        }
+      })
     }
     navigate("/tasks")
   })
@@ -36,43 +49,45 @@ function TaskFormPage() {
   }, [])
 
   return (
-    <div className='container'>
+    <div className='max-w-xl mx-auto'>
       <form onSubmit={submitData}>
-        <div className="form-group">
-          <label htmlFor="title">Titulo </label>
-          <input type="text" name="title" id="title" placeholder='Limpiar / Clean' className='form-control'
+          <input type="text" name="title" id="title" placeholder='Limpiar / Clean' className='bg-zinc-700 rounded-lg block w-full mb-3'
           {...register("title", {required: true}) }
           />
 
           {errors.title && <span>The title is required</span>}
-        </div>
         
-        <div className="form-group">
-          <label htmlFor="description">description</label>
-          <textarea name="description" id="description" cols="30" rows="10" placeholder='Lavar la loza / Clean the dishes ' className='form-control'
+          <textarea name="description" id="description" cols="30" rows="10" placeholder='Lavar la loza / Clean the dishes ' className='bg-zinc-700 rounded-lg block w-full mb-3'
             {...register("description", {required: true}) }
           ></textarea>
 
           {errors.description && <span>The description is required</span>}
 
-        </div>
 
-        <button>Save</button>
+        <button className='bg-indigo-500 p-3 rounded-lg block w-full mt-3'>Save</button>
       </form>
 
 
       {
         params.id && (
-          <button onClick={ async () => {
-            const acceptar = window.confirm('You sure to delete this task?')
+          <div className="flex justify-end">
+            <button className='bg-red-500 p-3 rounded-lg w-48 mt-3 ' onClick={ async () => {
+              const acceptar = window.confirm('You sure to delete this task?')
 
-            if(acceptar){
-              await deleteTask(params.id)
-              navigate("/tasks")
-            }
-          }}>
-            Delete
-          </button>
+                if(acceptar){
+                  await deleteTask(params.id)
+                  toast.success('Task succesfully deleted', {
+                    style: {
+                      background: "#101010",
+                      color: "#fff"
+                    }
+                  })
+                  navigate("/tasks")
+                }
+              }}>
+              Delete
+            </button>
+          </div>
         )
       }
